% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{friedman3}
\alias{friedman3}
\title{friedman3 data set}
\format{
A dummy data frame with 5 columns and 150 rows created using Benchmark Problem Friedman 3 by mlbench. The target feature is "y".
}
\source{
mlbench.friedman3(n = 150, sd = 3)
}
\usage{
friedman3
}
\description{
Data set to demonstrate regression task.
}
\keyword{datasets}
