% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{threenorm}
\alias{threenorm}
\title{threenorm data set}
\format{
A dummy data frame with 5 columns and 150 rows created using Threenorm Benchmark Problem by mlbench. The target feature is "classes".
}
\source{
mlbench.threenorm(150, d = 20)
}
\usage{
threenorm
}
\description{
Data set to demonstrate classification task.
}
\keyword{datasets}
